+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - Added possibility to override original variables `$display1-size` to `f$display4-size`.
 - Added possibility to override original variables `$display1-weight` to `$display4-weight`.
 - Fixed props `block` and `color` in Cards to be properly applied.
+- Updated Row Component added `noGutters`, `cssModules` and `tag` props.
 
 ## [1.4.1] - 2017-09-29